+++ .gitignore
@@ -9,6 +9,3 @@ bin/
 # Maven
 target
 dependency-reduced-pom.xml
-
-# Auto-generated files
-coverage
