+++ eslint-plugin/package.json
@@ -28,8 +28,8 @@
     "lint:js": "eslint .",
     "lint:fix": "eslint . --fix",
     "pack:sonar": "npm pkg set main=\"./dist/rules.js\" && mkdirp dist/pack && yarn pack -o dist/pack/creedengo-eslint-plugin.tgz && npm pkg set main=\"./lib/standalone.js\"",
-    "test": "mocha tests --recursive",
-    "test:cov": "nyc --reporter=lcov --reporter=text mocha tests --recursive",
+    "test": "node --test",
+    "test:cov": "node --test --experimental-test-coverage",
     "update:eslint-docs": "eslint-doc-generator"
   },
   "devDependencies": {
@@ -42,9 +42,8 @@
     "eslint-plugin-eslint-plugin": "^7.3.2",
     "eslint-plugin-license-header": "^0.9.0",
     "eslint-plugin-prettier": "^5.5.5",
+    "globals": "^17.6.0",
     "mkdirp": "^3.0.1",
-    "mocha": "^11.7.5",
-    "nyc": "^17.1.0",
     "prettier": "^3.8.3",
     "rimraf": "^6.1.3",
     "typescript": "~5.9.3"
