+++ eslint-plugin/tests/lib/rules/no-torch.test.js
@@ -22,52 +22,41 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const rule = require("../../../lib/rules/no-multiple-access-dom-element");
-const RuleTester = require("eslint").RuleTester;
+const rule = require("../../../lib/rules/no-torch");
+const { RuleTester } = require("eslint");
+const { describe, it } = require("node:test");
 
 //------------------------------------------------------------------------------
 // Tests
 //------------------------------------------------------------------------------
 
-const ruleTester = new RuleTester();
+const ruleTester = new RuleTester({
+  languageOptions: {
+    ecmaVersion: 6,
+    sourceType: "module",
+  },
+});
 const expectedError = {
-  messageId: "ShouldBeAssignToVariable",
+  messageId: "ShouldNotProgrammaticallyEnablingTorchMode",
 };
 
-ruleTester.run("no-multiple-access-dom-element", rule, {
+const tests = {
   valid: [
     `
-    var el1 = document.getElementById('block1').test;
-    var el2 = document.getElementById('block2').test
-    `,
-    `
-    var el1 = document.getElementsByClassName('block1').test;
-    var el2 = document.getElementsByClassName('block2').test
-    `,
-    `
-    function test() { var link = document.getElementsByTagName('a'); }
-    var link = document.getElementsByTagName('a');
-    `,
-    `
-    for (var i = 0; i < 10; i++) {
-      var test = document.getElementsByName("test" + i)[0].value;
-      var test2 = document.getElementsByName("test2" + i)[0].value;
-    }
+    import axios from 'axios';
     `,
   ],
 
   invalid: [
     {
-      code: "var el1 = document.getElementById('block1').test1; var el2 = document.getElementById('block1').test2",
-      errors: [expectedError],
-    },
-    {
-      code: "function test() {var el1 = document.getElementById('block1').test1; if(toto) { var el2 = document.getElementById('block1').test2 }}",
-      errors: [expectedError],
-    },
-    {
-      code: "if (true) { var card = document.querySelector('.card'); } else { var card = document.querySelector('.card'); }",
+      code: "import Torch from 'react-native-torch';",
       errors: [expectedError],
     },
   ],
+};
+
+describe("no-torch", () => {
+  it("no-torch", () => {
+    ruleTester.run("no-torch", rule, tests);
+  });
 });
