+++ sonar-project.properties
@@ -3,6 +3,6 @@ sonar.projectKey=green-code-initiative_ecoCode-linter
 sonar.projectName=creedengo - JavaScript language
 sonar.sources=eslint-plugin/lib/,eslint-plugin/dist/rules.js,sonar-plugin/src/main/java/
 sonar.tests=eslint-plugin/tests/,sonar-plugin/src/test/java/
-sonar.javascript.lcov.reportPaths=eslint-plugin/coverage/lcov.info
+sonar.javascript.lcov.reportPaths=eslint-plugin/lcov.info
 sonar.java.binaries=sonar-plugin/target/
 sonar.coverage.jacoco.xmlReportPaths=sonar-plugin/target/site/jacoco/jacoco.xml
